Update link for Linux Lazy Newb Pack
Changed the link from lazynewbpack.com as this domain has been hijacked and redirects to malware; removed references to Beautato pack. 

It seems beautato's pack hasn't been updated for 6 years, so I've included a link to the Wiki instead where there are 2 current forks of LNPs for Linux. 

If you wanted to pick one and direct link in the same way as the Mac and Windows links, the LinuxDwarfPack is currently the more up to date of the two, but I primarily wanted to remove the dodgy link.